Split web chat reconcile DB exporter
