+++ cmd/web-chat/app/debug_reconcile_schema.go
@@ -1,0 +1,50 @@
+package app
+
+import (
+	"context"
+	"database/sql"
+)
+
+func createDebugSQLiteSchema(ctx context.Context, db *sql.DB) error {
+	stmts := []string{
+		`CREATE TABLE meta (key TEXT PRIMARY KEY, value TEXT NOT NULL)`,
+		`CREATE TABLE backend_records (id INTEGER PRIMARY KEY AUTOINCREMENT, kind TEXT NOT NULL, ts TEXT, session_id TEXT, connection_id TEXT, ordinal INTEGER, raw_json TEXT NOT NULL)`,
+		`CREATE TABLE backend_pipeline (record_id INTEGER PRIMARY KEY REFERENCES backend_records(id), mode TEXT, event_name TEXT, event_type TEXT, event_appended INTEGER, view_ordinal INTEGER, timeline_cursor_advanced INTEGER, append_error TEXT, session_error TEXT, view_error TEXT, ui_projection_error TEXT, timeline_projection_error TEXT, apply_error TEXT, cursor_error TEXT, fanout_error TEXT)`,
+		`CREATE TABLE backend_pipeline_ui_events (id INTEGER PRIMARY KEY AUTOINCREMENT, record_id INTEGER NOT NULL REFERENCES backend_records(id), source TEXT NOT NULL, name TEXT, payload_type TEXT, payload_json TEXT)`,
+		`CREATE TABLE backend_pipeline_entities (id INTEGER PRIMARY KEY AUTOINCREMENT, record_id INTEGER NOT NULL REFERENCES backend_records(id), source TEXT NOT NULL, kind TEXT, entity_id TEXT, created_ordinal INTEGER, last_event_ordinal INTEGER, tombstone INTEGER, payload_type TEXT, payload_json TEXT)`,
+		`CREATE TABLE backend_transport (record_id INTEGER PRIMARY KEY REFERENCES backend_records(id), stage TEXT, direction TEXT, frame_type TEXT, event_name TEXT, payload_type TEXT, since_snapshot_ordinal INTEGER, snapshot_ordinal INTEGER, snapshot_entity_count INTEGER, fanout_event_count INTEGER, raw_bytes INTEGER, queue_len INTEGER, queue_cap INTEGER, error TEXT)`,
+		`CREATE TABLE backend_transport_snapshot_entities (id INTEGER PRIMARY KEY AUTOINCREMENT, record_id INTEGER NOT NULL REFERENCES backend_records(id), kind TEXT, entity_id TEXT, created_ordinal INTEGER, last_event_ordinal INTEGER, payload_type TEXT, tombstone INTEGER)`,
+		`CREATE TABLE geppetto_records (record_id INTEGER PRIMARY KEY REFERENCES backend_records(id), ts TEXT, provider TEXT, model TEXT, session_id TEXT, inference_id TEXT, turn_id TEXT, message_id TEXT, stage TEXT NOT NULL, event_type TEXT, info_message TEXT, response_id TEXT, item_id TEXT, output_index INTEGER, summary_index INTEGER, delta_len INTEGER, normalized_delta_len INTEGER, buffer_len INTEGER, error TEXT, object_json TEXT, event_json TEXT, metadata_json TEXT)`,
+		`CREATE TABLE geppetto_provider_events (record_id INTEGER PRIMARY KEY REFERENCES geppetto_records(record_id), provider_event_type TEXT, response_id TEXT, item_id TEXT, output_index INTEGER, summary_index INTEGER, object_json TEXT)`,
+		`CREATE TABLE geppetto_emitted_events (record_id INTEGER PRIMARY KEY REFERENCES geppetto_records(record_id), geppetto_event_type TEXT, info_message TEXT, response_id TEXT, item_id TEXT, output_index INTEGER, summary_index INTEGER, event_json TEXT, metadata_json TEXT)`,
+		`CREATE INDEX idx_geppetto_records_session_stage ON geppetto_records(session_id, stage)`,
+		`CREATE INDEX idx_geppetto_records_item ON geppetto_records(item_id)`,
+		`CREATE INDEX idx_geppetto_records_event_type ON geppetto_records(event_type)`,
+		`CREATE TABLE frontend_records (id INTEGER PRIMARY KEY, type TEXT NOT NULL, ts_ms INTEGER, ts_iso TEXT, session_id TEXT, ordinal INTEGER, raw_json TEXT NOT NULL)`,
+		`CREATE TABLE frontend_raw_ws (record_id INTEGER PRIMARY KEY REFERENCES frontend_records(id), size INTEGER, preview TEXT, raw TEXT)`,
+		`CREATE TABLE frontend_parsed_frames (record_id INTEGER PRIMARY KEY REFERENCES frontend_records(id), frame_type TEXT, name TEXT, payload_type TEXT, frame_json TEXT)`,
+		`CREATE TABLE frontend_snapshots (record_id INTEGER PRIMARY KEY REFERENCES frontend_records(id), entity_count INTEGER, dropped_count INTEGER)`,
+		`CREATE TABLE frontend_snapshot_entities (id INTEGER PRIMARY KEY AUTOINCREMENT, record_id INTEGER NOT NULL REFERENCES frontend_records(id), raw_kind TEXT, raw_id TEXT, mapped_kind TEXT, mapped_id TEXT, dropped INTEGER)`,
+		`CREATE TABLE frontend_ui_events (record_id INTEGER PRIMARY KEY REFERENCES frontend_records(id), name TEXT, message_id TEXT, mutation_json TEXT)`,
+		`CREATE TABLE frontend_lifecycle (record_id INTEGER PRIMARY KEY REFERENCES frontend_records(id), event TEXT)`,
+		`CREATE INDEX idx_backend_records_session_ordinal ON backend_records(session_id, ordinal)`,
+		`CREATE INDEX idx_backend_transport_stage ON backend_transport(stage)`,
+		`CREATE INDEX idx_frontend_records_session_ordinal ON frontend_records(session_id, ordinal)`,
+		`CREATE INDEX idx_frontend_records_type ON frontend_records(type)`,
+
+		// Timeline and turns tables — persisted data alongside event evidence.
+		`CREATE TABLE timeline_entities (id INTEGER PRIMARY KEY AUTOINCREMENT, kind TEXT NOT NULL, entity_id TEXT NOT NULL, created_ordinal INTEGER NOT NULL, last_event_ordinal INTEGER NOT NULL, tombstone INTEGER NOT NULL DEFAULT 0, payload_type TEXT, payload_json TEXT)`,
+		`CREATE INDEX idx_timeline_entities_kind ON timeline_entities(kind)`,
+		`CREATE INDEX idx_timeline_entities_entity_id ON timeline_entities(entity_id)`,
+
+		`CREATE TABLE turns (id INTEGER PRIMARY KEY AUTOINCREMENT, conv_id TEXT NOT NULL, session_id TEXT NOT NULL, turn_id TEXT NOT NULL, phase TEXT NOT NULL, runtime_key TEXT, inference_id TEXT, created_at_ms INTEGER, created_at TEXT, payload_json TEXT NOT NULL)`,
+		`CREATE INDEX idx_turns_session_id ON turns(session_id)`,
+		`CREATE INDEX idx_turns_conv_id ON turns(conv_id)`,
+	}
+	for _, stmt := range stmts {
+		if _, err := db.ExecContext(ctx, stmt); err != nil {
+			return err
+		}
+	}
+	return nil
+}
