+++ ttmp/2026/05/07/PINO-CODE-REVIEW--code-review-and-cleanup-guide-for-sessionstream-chatapp-webchat-integration/tasks.md
@@ -25,7 +25,7 @@
 
 - [ ] Split `pkg/chatapp/chat.go` into engine, runtime sink, projections, IDs, and demo inference files.
 - [x] Split `cmd/web-chat/app/debug_recorder.go` by pipeline, transport, and Geppetto record domains.
-- [ ] Extract `cmd/web-chat/app/debug_reconcile_db.go` into schema, inserts, frontend parsing, views, and provider adapters.
+- [x] Extract `cmd/web-chat/app/debug_reconcile_db.go` into schema, inserts, frontend parsing, views, and provider adapters.
 - [ ] Split `cmd/web-chat/web/src/ws/wsManager.ts` into transport client, hydration coordinator, event mapper, and debug hooks.
 - [ ] Add a generated or table-driven frontend mapping for typed `ReasoningUpdate`, `ToolCallUpdate`, and `ChatMessageUpdate` payloads.
 - [ ] Move app-local agent-mode plugin code out of `cmd/web-chat` or explicitly document it as example-only.
